requirements.txt
streamlit
swarms
python-dotenv